requirements.txt
python-dotenv
google-generativeai
pyPDF2
faiss-cpu
numpy
streamlit
-e .
